# docs/swagger.yaml
definitions:
...
  model.MySQLTime:
    properties:
      time.Time:
        type: string
    type: object
...
  model.User:
    properties:
      access_token:
        type: string
      active:
        type: boolean
      created_at:
        $ref: '#/definitions/model.MySQLTime'
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      role_id:
        type: integer
      updated_at:
        $ref: '#/definitions/model.MySQLTime'
    type: object
  model.UserResponse:
    properties:
      name:
        type: string
      role_id:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /login:
    post:
      description: Log in user with the provided credentials
      parameters:
      - description: User object that contains login credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      responses:
        "200":
          description: Login successful
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body" (when the request body does not contain
            valid JSON or is missing required fields)
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized" (when the provided credentials are incorrect)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to update access token" (when there is an error updating
            the access token in the database)
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
  /logout:
    post:
      description: Log out user by clearing the access token
      responses:
        "200":
          description: Logout successful
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized" (when the provided JWT token is invalid or missing)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error" (when there is a problem with the database)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Logout user
  /register:
    post:
      description: Register a new user with the provided information
      parameters:
      - description: User object that needs to be registered
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      responses:
        "200":
          description: Registration successful!
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Password should be at least 8 characters" (when the provided
            password is less than 8 characters)
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Email already registered" (when the provided email is already
            registered)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error" (when there is a problem with the database
            or password hashing)
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register new user
...
  /users:
    get:
      description: Get a list of users with limited information based on the user's
        role
      responses:
        "200":
          description: List of users
          schema:
            items:
              $ref: '#/definitions/model.UserResponse'
            type: array
        "401":
          description: Unauthorized" (when the provided JWT token is invalid or missing)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error" (when there is a problem with the database)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get users
    put:
      description: Update user data (name and password)
      parameters:
      - description: User object that contains updated user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      responses:
        "200":
          description: User data updated
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body" (when the request body does not contain
            valid JSON or is missing required fields)
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized" (when the provided JWT token is invalid or missing)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error" (when there is a problem with the database)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update user
  /users/{id}:
    delete:
      description: Delete a user by its ID (only accessible by admin)
      parameters:
      - description: User ID to be deleted
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: User successfully deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid userID" (when the provided user ID in the URL is not
            a valid integer)
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized" (when the provided JWT token is invalid or missing)
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied" (when the user does not have admin role)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found" (when the requested user ID does not exist
            in the database)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error" (when there is a problem with the database)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete user
    get:
      description: Get detailed information about a specific user (only accessible
        by admin)
      parameters:
      - description: User ID to be retrieved
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: User details
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Invalid userID" (when the provided user ID in the URL is not
            a valid integer)
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized" (when the provided JWT token is invalid or missing)
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied" (when the user does not have admin role)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found" (when the requested user ID does not exist
            in the database)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error" (when there is a problem with the database)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get user details
...
swagger: "2.0"